I moved the processText to the model instead of the controller.
This needed to be used in the controlelr, but is a job for the Chatbot to hold.

I also moved the JOptionPane from the Start method to the interactWithChatbot method.

I used the processText thing that I moved to Chatbot in the BotController to make the bot respond with what the user typed as a response.